Move locale-related models into wagtail.core.models.i18n
